% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.stanjm.R
\name{as.matrix.stanjm}
\alias{as.array.stanjm}
\alias{as.data.frame.stanjm}
\alias{as.matrix.stanjm}
\title{Extract the posterior sample}
\usage{
\method{as.matrix}{stanjm}(x, ..., pars = NULL, regex_pars = NULL)

\method{as.array}{stanjm}(x, ..., pars = NULL, regex_pars = NULL)

\method{as.data.frame}{stanjm}(x, ..., pars = NULL, regex_pars = NULL)
}
\arguments{
\item{x}{A fitted model object returned by the
\code{\link{stan_jm}} modelling function. See \code{\link{stanjm-object}}.}

\item{...}{Ignored.}

\item{pars}{An optional character vector of parameter names.}

\item{regex_pars}{An optional character vector of \link[=grep]{regular 
expressions} to use for parameter selection. \code{regex_pars} can be used
in place of \code{pars} or in addition to \code{pars}.}
}
\value{
A matrix, data.frame, or array, the dimensions of which depend on
  \code{pars} and \code{regex_pars} as well as the model.
}
\description{
For models fit using MCMC (\code{algorithm="sampling"}), the posterior sample
---the post-warmup draws from the posterior distribution--- can be extracted 
from a fitted model object as a matrix, data frame, or array. The 
\code{as.matrix} and \code{as.data.frame} methods merge all chains together, 
whereas the \code{as.array} method keeps the chains separate.
}
\seealso{
\code{\link{stanjm-methods}}
}

