% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplejm.R
\name{examplejm}
\alias{examplejm}
\title{Example joint longitudinal and time-to-event model}
\format{Calling \code{example("examplejm")} will run the model in the 
  \strong{Examples} section below, and the resulting stanjm object will 
  then be available in the global environment. The example uses
  the Mayo Clinic's primary biliary cirrhosis dataset, however, the example
  only uses a small subset of the patients so that the model runs
  quickly. Type \code{help("pbc-datasets", package = "rstanjm")} for a  
  brief description of the data.\cr
  \cr
  Also, note that by default rstanjm fits models using a single MCMC chain. 
  However, the preference is to run multiple MCMC chains to help assess
  convergence and also increase the effective sample size. The number of 
  chains can be increased by simply specifying the \code{chains} argument
  when fitting the model. If you have a multicore CPU with excess RAM, you 
  can fit multiple MCMC chains in parallel by setting the \code{cores} 
  argument equal to the number of chains being executed. This is preferable 
  since then the multiple chains will be run in parallel, thereby saving on 
  computation time.}
\description{
A model for use in the \pkg{rstanjm} examples.
}
\examples{

  examplejm <- 
     stan_jm(formulaLong = logBili ~ year + (1 | id), 
             dataLong = pbcLong_subset,
             formulaEvent = Surv(futimeYears, death) ~ sex + trt, 
             dataEvent = pbcSurv_subset,
             time_var = "year",
             chains = 1, cores = 1, seed = 12345)


}

