% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanjm-methods.R
\name{family.stanjm}
\alias{family.stanjm}
\title{family method for stanjm objects}
\usage{
\method{family}{stanjm}(object, ...)
}
\arguments{
\item{object, ...}{See \code{\link[stats]{family}}.}
}
\description{
family method for stanjm objects
}
\keyword{internal}

