% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_y}
\alias{get_x}
\alias{get_y}
\alias{get_z}
\title{Extract X, Y or Z from a stanjm object}
\usage{
get_y(object)

get_x(object)

get_z(object)
}
\arguments{
\item{object}{A fitted model object returned by the
\code{\link{stan_jm}} modelling function. See \code{\link{stanjm-object}}.}
}
\value{
For \code{get_x} and \code{get_z}, list of matrices. For \code{get_y}, either
  a vector or a matrix, depending on how the response variable was specified.
}
\description{
Extract X, Y or Z from a stanjm object
}
\keyword{internal}

