% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanjm-methods.R
\name{log_lik}
\alias{log_lik}
\title{Extract pointwise log-likelihood matrix}
\usage{
log_lik(object, ...)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Arguments to methods. For example the
\code{\link[=stanjm-methods]{stanjm}} method accepts the arguments
\code{newdataLong} and \code{newdataEvent}.}
}
\value{
The \eqn{S} by \eqn{N} pointwise log-likelihood matrix,
  where \eqn{S} is the size of the posterior sample and \eqn{N} is the number
  of individuals in the fitted model. The likelihood for a single individual 
  is therefore the sum of the likelihood contributions from their observed
  longitudinal measurements and their event time data.
}
\description{
Extract pointwise log-likelihood matrix
}
\seealso{
\code{\link{log_lik.stanjm}}
}
\keyword{internal}

