% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanjm-methods.R
\name{model.frame.stanjm}
\alias{model.frame.stanjm}
\title{model.frame method for stanjm objects}
\usage{
\method{model.frame}{stanjm}(formula, fixed.only = FALSE, ...)
}
\arguments{
\item{formula, ...}{See \code{\link[stats]{model.frame}}.}

\item{fixed.only}{See \code{\link[lme4]{model.frame.merMod}}.}
}
\description{
model.frame method for stanjm objects
}
\keyword{internal}

