% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanjm-methods.R
\name{model.matrix.stanjm}
\alias{model.matrix.stanjm}
\title{model.matrix method for stanjm objects}
\usage{
\method{model.matrix}{stanjm}(object, ...)
}
\arguments{
\item{object, ...}{See \code{\link[stats]{model.matrix}}.}
}
\description{
model.matrix method for stanjm objects
}
\keyword{internal}

