% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predict.stanjm.R
\name{plot.predict.stanjm}
\alias{plot.predict.stanjm}
\title{Plot estimated subject-specific longitudinal trajectory}
\usage{
\method{plot}{predict.stanjm}(x, ids = NULL, limits = c("ci", "pi", "none"),
  xlab = NULL, ylab = NULL, abline = FALSE, plot_observed = FALSE,
  facet_scales = "free_x", ci_geom_args = NULL, ...)
}
\arguments{
\item{x}{A data frame and x of class \code{predict.stanjm}
returned by a call to the function \code{\link{posterior_traj}}.
The x contains point estimates and uncertainty interval limits
for fitted values of the longitudinal response.}

\item{ids}{An optional vector providing a subset
of IDs for whom the predicted curves should be plotted.}

\item{limits}{A quoted character string specifying the type of limits to
include in the plot. Can be one of: \code{"ci"} for the Bayesian
posterior uncertainty interval (often known as a credible interval);
\code{"pi"} for the prediction interval; or \code{"none"} for no
interval limits.}

\item{xlab, ylab}{An optional axis label passed to 
\code{\link[ggplot2]{labs}}.}

\item{abline}{A logical. If \code{TRUE} then a vertical dashed line
is added to the plot indicating the event or censoring time for
the individual. Can only be used if each plot within the figure
is for a single individual.}

\item{plot_observed}{A logical. If \code{TRUE} then the observed
longitudinal measurements are overlaid on the plot.}

\item{facet_scales}{A character string passed to the \code{scales}
argument of \code{\link[ggplot2]{facet_wrap}} when plotting the 
longitudinal trajectory for more than one individual.}

\item{ci_geom_args}{Optional arguments passed to 
\code{\link[ggplot2]{geom_ribbon}} and used to control features
of the plotted interval limits.}

\item{...}{Optional arguments passed to 
\code{\link[ggplot2]{geom_smooth}} and used to control features
of the plotted trajectory.}
}
\value{
A \code{ggplot} x, also of class \code{plot.predict.stanjm}.
  This x can be further customised using the \pkg{ggplot2} package.
  It can also be passed to the function \code{\link{plot_stack}}.
}
\description{
Various plots comparing the observed outcome variable from one
of the longitudinal submodels \eqn{y} to simulated 
datasets \eqn{y^{rep}}{yrep} from the posterior predictive distribution.
This function is modelled on the \code{\link[rstanarm]{pp_check}} function
from the \pkg{rstanarm} package.
}
\examples{

}
\seealso{
\code{\link{posterior_traj}}, \code{\link{plot_stack}}
}

