% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.stanjm}
\alias{plot.stanjm}
\title{Plot method for stanjm objects}
\usage{
\method{plot}{stanjm}(x, plotfun = NULL, pars = NULL, regex_pars = NULL,
  ...)
}
\arguments{
\item{x}{A fitted model object returned by the
\code{\link{stan_jm}} modelling function. See \code{\link{stanjm-object}}.}

\item{plotfun}{A character string naming the plotting function to apply to 
the stanjm object. See \code{\link[rstanarm]{rstanarm-plots}} for the names and
descriptions. Also see the \strong{Examples} section below. \code{plotfun} can be
either the full name of the plotting function (e.g. \code{"stan_hist"}) or
can be abbreviated to the part of the name following the underscore (e.g. 
\code{"hist"}). The default plot displays intervals and point estimates for 
the parameters.}

\item{pars}{An optional character vector specifying a subset of parameters to
display. Parameters can be specified by name or several shortcuts can be 
used. 
Using \code{pars = "long"} will display the 
parameter estimates for the longitudinal submodels only (excluding random
effects, but including dispersion parameters).
Using \code{pars = "event"} will display the 
parameter estimates for the event submodel only, including any association
parameters. 
Using \code{pars = "assoc"} will display only the 
association parameters. 
Using \code{pars = "fixef"} will display all fixed effects, but not
the random effects or the additional parameters such as dispersion, etc. 
Using \code{pars = "beta"} will display only the 
fixed effect regression coefficients (excluding the intercept terms).
Using \code{pars = "alpha"} will display only the 
fixed effect intercept terms.
The estimates for the random effects can be selected using \code{pars = "b"}
or \code{pars = "varying"}.
If both \code{pars} and \code{regex_pars} are set to \code{NULL} then all 
parameters are selected, including the log posterior. 
If both \code{pars} and \code{regex_pars} are set to \code{NULL} then all 
fixed effect regression coefficients are selected (with the exception
of the intercept term if a Weibull baseline hazard was estimated).}

\item{regex_pars}{An optional character vector of \link[=grep]{regular 
expressions} to use for parameter selection. \code{regex_pars} can be used
in place of \code{pars} or in addition to \code{pars}.}

\item{...}{Additional arguments to pass to \code{plotfun} (see
\code{\link[rstanarm]{rstanarm-plots}}).}
}
\value{
A ggplot object (or several) that can be further 
  customized using the \pkg{ggplot2} package.
}
\description{
For models fit using \code{\link{stan_jm}}, there are a variety of 
diagnostic and inference related plots that can be generated. The 
user can specify which type of plot to display via the 
\code{plotfun} argument.
}
\examples{

}
\seealso{
\code{\link[rstanarm]{rstanarm-plots}} for details on the individual plotting
  functions.
}

