% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stack.R
\name{plot_stack}
\alias{plot_stack}
\title{Combine plots of estimated longitudinal trajectories and survival function}
\usage{
plot_stack(yplot, survplot)
}
\arguments{
\item{yplot}{An object of class \code{plot.predict.stanjm}, returned by a 
call to the plotting method for objects of class 
\code{predict.stanjm}. If there is more than one longitudinal outcome,
then a list of such objects can be provided.}

\item{survplot}{An object of class \code{plot.survfit.stanjm}, returned by a 
call to the plotting method for objects of class 
\code{survfit.stanjm}.}
}
\value{
A single \code{ggplot} object that includes plots of the estimated 
  subject-specific longitudinal trajectories stacked on top of their 
  associated subject-specific survival curves.
}
\description{
This function takes plots of estimated subject-specific longitudinal 
trajectories and the estimated subject-specific survival function and 
combines them into a single figure.
}

