% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_interval.R
\name{posterior_interval}
\alias{posterior_interval}
\title{Posterior uncertainty intervals}
\usage{
posterior_interval(object, ...)
}
\arguments{
\item{object}{A fitted model object returned by the
\code{\link{stan_jm}} modelling function. See \code{\link{stanjm-object}}.}

\item{...}{Arguments passed to the \code{\link[rstanarm]{posterior_interval}} 
function of the \pkg{rstanarm} package.}
}
\value{
A matrix with two columns and as many rows as model parameters (or 
  the subset of parameters specified by \code{pars} and/or 
  \code{regex_pars}). For a given value of \code{prob}, \eqn{p}, the columns 
  correspond to the lower and upper \eqn{100p}\% interval limits and have the
  names \eqn{100\alpha/2}\% and \eqn{100(1 - \alpha/2)}\%, where \eqn{\alpha 
  = 1-p}. For example, if \code{prob=0.9} is specified (a \eqn{90}\%
  interval), then the column names will be \code{"5\%"} and \code{"95\%"},
  respectively.
}
\description{
The \code{posterior_interval} function computes Bayesian posterior  
uncertainty intervals for the model parameters, via a call to the 
\code{posterior_interval} function in the \pkg{rstanarm} package.
See \code{\link[rstanarm]{posterior_interval}} for details.
}
\examples{
if (!exists("examplejm")) example(examplejm)
posterior_interval(examplejm)
posterior_interval(examplejm, regex_pars = "Long1")
posterior_interval(examplejm, regex_pars = c("Event", "Assoc"))
posterior_interval(examplejm, regexpars = "b[", prob = 0.95)
posterior_interval(examplejm, pars = "Long1|(Intercept)", prob = 0.5)

}
\references{
Gelman, A. and Carlin, J. (2014). Beyond power calculations:
assessing Type S (sign) and Type M (magnitude) errors. \emph{Perspectives on
Psychological Science}. 9(6), 641--51.

Morey, R. D., Hoekstra, R., Rouder, J., Lee, M. D., and Wagenmakers, E. 
(2015). The fallacy of placing confidence in confidence intervals. 
\emph{Psychonomic Bulletin & Review}. 1--21.
}

