% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{print.stanjm}
\alias{print.stanjm}
\title{Print method for stanjm objects}
\usage{
\method{print}{stanjm}(x, digits = 3, ...)
}
\arguments{
\item{x}{A fitted model object returned by the
\code{\link{stan_jm}} modelling function. See \code{\link{stanjm-object}}.}

\item{digits}{Number of digits to use for formatting numbers.}

\item{...}{Ignored.}
}
\value{
Returns \code{x}, invisibly.
}
\description{
The \code{print} method for \code{\link[=stanjm-object]{stanjm}}
objects displays a compact summary of a joint model fitted using the 
\code{\link{stan_jm}} modelling function. For more detailed summary statistics 
and diagnostics use the \code{\link[=summary.stanjm]{summary}} method.
}
\details{
The output from the \code{print} method includes very brief summary statistics 
for the fixed effects parameter estimates (described below, and which are 
displayed separately for each longitudinal or event submodel), the additional
parameters such as the residual error standard deviation, dispersion parameter, 
or the Weibull shape parameter or baseline hazard coefficients. Also displayed 
are the standard deviation and correlation parameter estimates for the random effects.  

\subsection{Point estimates}{
Point estimates are medians computed from the posterior sample generated 
via the MCMC simulations. The point estimates reported are the same as the 
values returned by the \code{\link[=coef.stanjm]{coef}} method.
}
\subsection{Uncertainty estimates}{
The standard deviations (labeled MAD_SD in the print output) are computed from
the same posterior sample used to calculate the point estimates and are proportional 
to the median absolute deviation (\code{\link[stats]{mad}}) from the median. 
Compared to the raw posterior standard deviation, the MAD_SD will be more 
robust for long-tailed distributions. These are the same as the values 
returned by \code{\link[=se.stanjm]{se}}.
}
}
\seealso{
\code{\link{summary.stanjm}}, \code{\link{stanjm-methods}}
}

