% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priorAssoc_options}
\alias{priorEvent_options}
\alias{priorLong_options}
\alias{priors}
\title{Additional options related to prior distributions for fitting Bayesian 
joint models via Stan}
\usage{
priorLong_options(prior_scale_for_dispersion = 5, min_prior_scale = 1e-12,
  scaled = TRUE)

priorEvent_options(prior_scale_for_weibull = 5, min_prior_scale = 1e-12,
  scaled = TRUE)

priorAssoc_options(min_prior_scale = 1e-12)
}
\arguments{
\item{prior_scale_for_dispersion}{Prior scale for the standard error of the 
regression in Gaussian models, which is given a half-Cauchy prior truncated
at zero.}

\item{min_prior_scale}{Minimum prior scale for the intercept and 
coefficients.}

\item{scaled}{A logical scalar, defaulting to \code{TRUE}. If \code{TRUE} the
\code{prior_scale} is further scaled by the range of the predictor if the 
predictor has exactly two unique values and scaled by twice the standard
deviation of the predictor if it has more than two unique values.}

\item{prior_scale_for_weibull}{Prior scale for the shape parameter of the 
Weibull distribution for the baseline hazard in \code{\link{stan_glm}}, 
which is given a half-Cauchy truncated at zero.}
}
\value{
A named list to be used internally by the \code{link{stan_jm}}
  modelling function.
}
\description{
These functions are used to specify additional options related to
prior distributions when fitting a joint model using the \code{link{stan_jm}} 
function, which is the main modeling function in the \pkg{rstanjm} package. 
However, the main functions used for specifying prior distributions for 
parameters in the joint model are called from the \pkg{rstanarm} package; 
see \code{link[rstanarm]{priors}} for further details.
}
\details{
The functions shown here are used to generate additional options
  related to prior distributions for parameters in a joint longitudinal 
  and time-to-event model, estimated using the \code{link{stan_jm}}
  modelling function. The functions are used to specify options separately
  for the longitudinal submodel(s), the event submodel, and the association
  parameters.
  
  The prior distributions for the regression coefficients in each of the 
  submodels as well as the prior distribution on the covariance matrix for
  the group-specific parameters (or subject-specific random effects) are
  implemented via the \pkg{rstanarm} package; see \code{link[rstanarm]{priors}} 
  for details.
}

