% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanjm-package.R
\docType{package}
\name{rstanjm-package}
\alias{rstanjm}
\alias{rstanjm-package}
\title{Bayesian joint longitudinal and time-to-event models via Stan}
\description{
The \pkg{rstanjm} package allows the user to fit "joint models" for 
longitudinal and time-to-event data (also known as shared parameter models) under
a Bayesian framework. Both 
univariate (one longitudinal marker) and multivariate (more than one longitudinal 
marker) joint models are allowed. Both continuous and non-continuous (e.g.
binary or count data) longitudinal markers can be accomodated through a range
of possible link functions and error distributions. Multi-level clustered data 
are allowed (e.g. patients within clinics), provided that the
individual (e.g. patient) is the lowest level of clustering. Competing risk events 
are not currenly allowed. \cr
\cr  
For each longitudinal submodel, a generalised linear mixed model is assumed with 
one of the family and link combinations allowed by the \pkg{\link{lme4}} package. \cr
\cr
If a multivariate joint model is specified (i.e. more than one longitudinal marker)
then dependence between the different longitudinal markers is captured through 
correlated random effects; that is, a multivariate normal distribution with an 
unstructured correlation matrix is assumed for all random effects within a given 
clustering level (e.g. individual-level random effects). \cr
\cr
If a model with multi-level clustering (i.e. clustering beyond the level of the 
individual) is specified, then a different correlation matrix is estimated for 
each clustering level. \cr
\cr
For the time-to-event submodel a parametric proportional hazards model is
assumed. The baseline hazard can be estimated using either a Weibull
distribution or approximated using restricted cubic splines.
Time-varying covariates are allowed in both the longitudinal and event 
submodels. Competing risk events are not currently allowed. The association 
structure for the joint model can be based on any of the following 
parameterisations: 
  current value of the linear predictor in the longitudinal submodel(s);
  current expected value in the longitudinal submodel(s);
  first derivative (slope) of the linear predictor in the longitudinal 
    submodel(s); 
  shared random effects;
  no association structure (equivalent to fitting separate longitudinal 
    and survival models). \cr
\cr
Estimation of the joint model is under a Bayesian framework using Markov chain 
Monte Carlo (MCMC) methods, with the underlying estimation implemented using 
Stan (\url{http://mc-stan.org/}). A variety of prior distributions are available
for the regression coefficients, including normal, Cauchy and t distributions, 
as well as shrinkage priors. The prior distribution for the covariance matrix 
of the individual-level (or other clustering-level) random effects is based on 
a novel decomposition. All prior distributions are implemented via the 
\pkg{rstanarm} package; see \code{\link[rstanarm]{priors}} for details.
}
\author{
Sam Brilleman <\email{sam.brilleman@monash.edu}>. \cr
  \cr
  Whilst acknowledging this package draws heavily on code from the 
  \pkg{\link{rstanarm}} package, of which I am not an author.
}
\references{
Stan Development Team. (2015). \emph{Stan Modeling Language Users Guide and
Reference Manual.} \url{http://mc-stan.org/documentation/}
}
\seealso{
\code{\link{stan_jm}} for the main modelling function, as well as
  \code{\link{stanjm-object}}, \code{\link{stanjm-methods}},
  \code{\link{posterior_predict}}, \code{\link{posterior_survfit}}.

Also see \url{http://mc-stan.org/} for more information on the Stan C++ 
  package used for model fitting.
}

