% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_testing.R
\name{simulated-datasets}
\alias{simdata_jm_cvassoc}
\alias{simdata_jm_cvassoc_id}
\alias{simdata_mjm_cvassoc}
\alias{simdata_mjm_cvassoc_id}
\alias{simulated-datasets}
\title{Simulated datasets used for \pkg{rstanjm} tests}
\format{\describe{
The simulated datasets each include 250 individuals. A binary
treatment code is assigned to each individual, with 
probability of 0.5. 

The datasets with \code{jm} in the name use a 
univariate joint model in the data generating process,
whilst the datasets with \code{mjm} in the name use
a bivariate joint model for the data generating process.   
  
The datasets without the \code{_id} suffix contain both  
longitudinal and event time data (they include one or more 
rows per individual). The datasets with the \code{_id} suffix 
contain event time data only (they include only one row per 
individual). \cr
\cr
The variable definitions are as follows: \cr
  \item{\code{id}}{numeric ID unique to each individual}
  \item{\code{trt}}{binary treatment variable taking the 
    values \code{0} or \code{1}}   
  \item{\code{t0}}{start of the time interval, taken to 
    be the measurement time for the longitudinal marker(s)
    \code{y1}, \code{y2}, etc.}
  \item{\code{t}}{end of the time interval, taken to be the 
    event time for the event indicator \code{d}}
  \item{\code{y1}, \code{y2}, ...}{the longitudinal 
    response of type gaussian (\code{_gaus}), poisson 
    (\code{_pois}), or bernoulli (\code{_bern}).}
  \item{\code{d}}{the event indicator}
}}
\description{
Simulated datasets used for running tests on the \pkg{rstanjm} package.
}
\keyword{internal}

