% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanjm-methods.R
\name{stanjm-methods}
\alias{VarCorr}
\alias{VarCorr.stanjm}
\alias{coef.stanjm}
\alias{fitted.stanjm}
\alias{fixef}
\alias{fixef.stanjm}
\alias{formula.stanjm}
\alias{log_lik.stanjm}
\alias{ngrps}
\alias{ngrps.stanjm}
\alias{ranef}
\alias{ranef.stanjm}
\alias{residuals.stanjm}
\alias{se.stanjm}
\alias{sigma}
\alias{sigma.stanjm}
\alias{stanjm-methods}
\alias{update.stanjm}
\alias{vcov.stanjm}
\title{Methods for stanjm objects}
\usage{
\method{coef}{stanjm}(object, ...)

\method{fitted}{stanjm}(object, ...)

\method{log_lik}{stanjm}(object, newdata = NULL, ...)

\method{residuals}{stanjm}(object, ...)

\method{se}{stanjm}(object, ...)

\method{formula}{stanjm}(x, fixed.only = FALSE, random.only = FALSE, ...)

\method{update}{stanjm}(object, formulaLong., formulaEvent., ...,
  evaluate = TRUE)

\method{vcov}{stanjm}(object, correlation = FALSE, ...)

\method{fixef}{stanjm}(object, remove_stub = TRUE, ...)

\method{ngrps}{stanjm}(object, ...)

\method{ranef}{stanjm}(object, ...)

\method{sigma}{stanjm}(object, ...)

\method{VarCorr}{stanjm}(x, sigma = 1, ...)
}
\arguments{
\item{...}{Ignored, except by the \code{update} method. See
\code{\link{update}}.}

\item{newdata}{For \code{log_lik}, an optional data frame of new data (e.g. 
holdout data) to use when evaluating the log-likelihood. See the 
description of \code{newdata} for \code{\link{posterior_predict}}.}

\item{fixed.only}{A logical specifying whether to only retain the fixed effect
part of the longitudinal submodel formulas}

\item{random.only}{A logical specifying whether to only retain the random effect
part of the longitudinal submodel formulas}

\item{correlation}{For \code{vcov}, if \code{FALSE} (the default) the
covariance matrix is returned. If \code{TRUE}, the correlation matrix is
returned instead.}

\item{remove_stub}{Logical specifying whether or not to remove the string
identifying the submodel from each of the coefficient names}

\item{sigma}{Ignored (included for compatibility with
\code{\link[nlme]{VarCorr}}).}

\item{formula., evaluate}{See \code{\link[stats]{update}}.}

\item{remove_stub}{Logical specifying whether or not to remove the string
identifying the submodel from each of the coefficient names}
}
\description{
S3 methods for \link[=stanjm-object]{stanjm} objects. There are also 
several methods (listed in See Also, below) with their own individual help
pages.
}
\details{
Most of these methods are similar to the methods defined for objects
  of class 'lm', 'glm', 'glmer', etc. However there are a few exceptions:
  
\describe{
#' \item{\code{coef}}{
Returns a list equal to the length of the number of submodels. The first
elements of the list are the coefficients from each of the fitted longitudinal
submodels and are the same layout as those returned by \code{coef} method of the
\pkg{lme4} package, that is, the sum of the random and fixed effects coefficients 
for each explanatory variable for each level of each grouping factor. The final
element of the returned list is a vector of fixed effect coefficients from the
event submodel. 
}

\item{\code{confint}}{
For models fit using optimization, confidence intervals are returned via a
call to \code{\link[stats]{confint.default}}. If \code{algorithm} is
\code{"sampling"}, \code{"meanfield"}, or \code{"fullrank"}, the
\code{\link{posterior_interval}} function should be used to compute Bayesian
uncertainty intervals.
}

\item{\code{log_lik}}{
For models fit using MCMC only, the \code{log_lik}
function returns the \eqn{S} by \eqn{N} pointwise log-likelihood matrix,
where \eqn{S} is the size of the posterior sample and \eqn{N} is the number
of data points. Note: we use \code{log_lik} rather than defining a
\code{\link[stats]{logLik}} method because (in addition to the conceptual
difference) the documentation for \code{logLik} states that the return value
will be a single number, whereas we return a matrix.
}

\item{\code{residuals}}{
Residuals are \emph{always} of type \code{"response"} (not \code{"deviance"}
residuals or any other type). However, in the case of \code{\link{stan_polr}}
with more than two response categories, the residuals are the difference 
between the latent utility and its linear predictor.
}
\item{\code{coef}}{
Medians are used for point estimates. See the \emph{Point estimates} section
in \code{\link{print.stanjm}} for more details.
}
\item{\code{se}}{
The \code{se} function returns standard errors based on 
\code{\link{mad}}. See the \emph{Uncertainty estimates} section in
\code{\link{print.stanjm}} for more details.
}
}
}
\seealso{
Other S3 methods for stanjm objects, which have separate documentation, 
including \code{\link{as.matrix.stanjm}}, \code{\link{plot.stanjm}}, 
\code{\link{predict.stanjm}}, \code{\link{print.stanjm}}, and
\code{\link{summary.stanjm}}.

\code{\link{posterior_interval}} and \code{\link{posterior_predict}} for 
alternatives to \code{confint} and \code{predict} for models fit using MCMC 
or variational approximation.
}

