% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanjm-methods.R
\name{terms.stanjm}
\alias{terms.stanjm}
\title{terms method for stanjm objects}
\usage{
\method{terms}{stanjm}(x, ..., fixed.only = TRUE, random.only = FALSE)
}
\arguments{
\item{x, fixed.only, random.only, ...}{See lme4:::terms.merMod.}
}
\description{
terms method for stanjm objects
}
\keyword{internal}

