% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_model.R
\name{examplejm}
\alias{examplejm}
\title{Example joint longitudinal and time-to-event model}
\format{Calling \code{example("examplejm")} will run the model in the 
  Examples section, below, and the resulting stanjm object will then be
  available in the global environment. The \code{cores} argument is 
  optional and on a multicore system, the user may well want to set 
  that equal to the number of chains being executed.}
\description{
A model for use in \pkg{rstanjm} examples.
}
\examples{
examplejm <- 
  stan_jm(formulaLong = logBili ~ year + (1 | id), 
          dataLong = pbcLong_subset,
          formulaEvent = Surv(futimeYears, death) ~ sex + trt, 
          dataEvent = pbcSurv_subset,
          time_var = "year",
          chains = 1, cores = 1, seed = 12345)

}

