% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanjm-methods.R
\name{log_lik}
\alias{log_lik}
\title{Extract pointwise log-likelihood matrix}
\usage{
log_lik(object, ...)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Arguments to methods. For example the
\code{\link[=stanjm-methods]{stanjm}} method accepts the argument
\code{newdata}.}
}
\value{
Pointwise log-likelihood matrix.
}
\description{
Extract pointwise log-likelihood matrix
}
\seealso{
\code{\link{log_lik.stanjm}}
}
\keyword{internal}

