% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_survfit.R
\name{posterior_survfit}
\alias{posterior_survfit}
\title{Estimate the marginal or subject-specific survival function}
\usage{
posterior_survfit(object, newdataEvent = NULL, newdataLong = NULL, ids,
  times = NULL, limits = c(0.025, 0.975), condition = TRUE,
  extrapolate = TRUE, extrapolate_args = list(dist = NULL, prop = 0.5,
  increments = 15), marginalised = FALSE, draws = NULL, fun = NULL,
  seed = NULL, ...)
}
\arguments{
\item{object}{A fitted model object returned by the
\code{\link{stan_jm}} modelling function. See \code{\link{stanjm-object}}.}

\item{newdataLong, newdataEvent}{Optionally, new data frames in which to look 
for variables with which to predict. The data should be provided in the same
format as for the original call to the estimation function 
\code{\link{stan_jm}}.}

\item{ids}{An optional character vector containing the IDs of individuals for 
whom the estimated survival probabilites should be obtained. This defaults 
to \code{NULL} which returns predictions for all individuals in the original 
model if \code{newdataEvent} is \code{NULL}, or otherwise all individuals 
in \code{newdataEvent}.}

\item{times}{An optional scalar or numeric vector specifying the values of 
\code{time_var} in the original model at which to obtain the estimated 
survival probabilities. If not specified, and \code{marginalised = FALSE}, 
then the default behaviour is to use the latest observation time for each 
individual (taken to be the event or censoring time if \code{newdata} is 
not specified, or the lastest value of \code{time_var} if \code{newdata} 
is specified) and extrapolate forward from there. If not specified, and
\code{marginalised = TRUE}, then the default behaviour is to calculate
the marginal survival probability at \code{n_increments} time points between
baseline and the latest observation or event/censoring time for all individuals.}

\item{limits}{A numeric vector of length 2 specifying the limits to use
for the credible interval.}

\item{condition}{A logical specifying whether the estimated subject-specific
survival probabilities at time \code{t} should be conditioned on 
survival up to a fixed time point \code{u}. The default is to condition
on the latest observation time for each individual (taken to be the event 
or censoring time if \code{newdata} is not specified, or the lastest value of 
\code{time_var} if \code{newdata} is specified). It cannot
be set to \code{TRUE} if \code{marginalised} is set to \code{TRUE}.}

\item{extrapolate}{A logical specifying whether to extrapolate the estimated 
survival function beyond the times specified in the \code{times} argument
If \code{TRUE} then the extrapolation can be further controlled using
the \code{extrapolate_args} argument.}

\item{extrapolate_args}{A named list with parameters controlling extrapolation 
of the estimated survival function when \code{extrapolate = TRUE}. The list
can contain the following named elements: 
\code{dist}, a positive scalar 
specifying the amount of time across which to forecast the estimated survival
function; 
\code{prop}, a positive scalar between 0 and 1 specifying the 
amount of time across which to forecast the estimated survival function but
represented as a proportion of the total observed follow up time for each
individual;
\code{increments}, a positive integer specifying the number of increments in 
time at which to calculate the forecasted survival probabilities. 
See the \strong{Examples} below.}

\item{marginalised}{A logical specifying whether the estimated 
subject-specific survival probabilities should be averaged (marginalised)
across all individuals for whom the predictions were obtained. This can be
used to obtain an estimate of the marginal survival probability. It cannot
be set to \code{TRUE} if \code{condition} is set to \code{TRUE}.}

\item{draws}{An integer indicating the number of MCMC draws to return. The default
and maximum number of draws is the size of the posterior sample.}

\item{fun}{An optional function to apply to the results. \code{fun} is found 
by a call to \code{\link{match.fun}} and so can be specified as a function
object, a string naming a function, etc.}

\item{seed}{An optional \code{\link[=set.seed]{seed}} to use.}

\item{...}{Currently unused.}
}
\value{
A named list with two elements, 'times' and 'survprobs'. The former
  is a list of times at which the survival probabilities are calculated. 
  The latter is a list containing the correponding survival probabilities.
  Each element of 'survprobs' contains a \code{draws} by \code{levels(ids)} 
  matrix of estimated survival probabilities. Each row of the matrix is a
  vector of predictions generated using a single draw of the model parameters
  from the posterior distribution.
}
\description{
The posterior predictive distribution is the distribution of the outcome 
implied by the model after using the observed data to update our beliefs 
about the unknown parameters in the model. Simulating data from the posterior
predictive distribution using the observed predictors is useful for checking 
the fit of the model. Drawing from the posterior predictive distribution at 
interesting values of the predictors also lets us visualize how a 
manipulation of a predictor affects (a function of) the outcome(s). With new 
observations of predictor variables we can use the posterior predictive 
distribution to generate predicted outcomes.
}
\examples{

}
\seealso{
\code{\link{ps_check}} for for graphical checks of the estimated 
  marginal survival function, and \code{\link{posterior_predict}} for 
  drawing from the posterior predictive distribution for the longitudinal 
  submodel.
}

