% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanjm-package.R
\docType{package}
\name{rstanjm-package}
\alias{rstanjm}
\alias{rstanjm-package}
\title{Bayesian joint longitudinal and time-to-event models via Stan}
\description{
This package allows the user to fit joint models for longitudinal 
and time-to-event data (also known as shared parameter models). Both 
univariate (one longitudinal marker) and multivariate (more than one longitudinal 
marker) joint models are allowed. Both continuous and non-continuous (e.g.
binary or count data) longitudinal markers can be accomodated through a range
of possible link functions and error distributions. Multi-level clustered data 
can be accomodated (for example, patients within clinics). Competing risk events 
are not currenly allowed. \cr
\cr  
For each longitudinal submodel, a generalised linear mixed model is assumed with 
one of the family and link combinations allowed by the \pkg{\link{lme4}} package.
If a multivariate joint model is specified (i.e. more than one longitudinal marker)
then dependence between the different longitudinal markers is captured through 
correlated random effects; that is, a multivariate normal distribution with an 
unstructured correlation matrix is assumed for all random effects within a given 
clustering level (e.g. individual-level random effects). \cr
\cr
For the time-to-event submodel the only option currently available is a Weibull 
proportional hazards model. Time-varying covariates are allowed in both the 
longitudinal and event submodels. Competing risk events are not 
currently allowed. The association structure for the joint model 
can be based on any of the following parameterisations: 
   current expected value in the longitudinal submodel(s);
   current value of the linear predictor in the longitudinal submodel(s);
   first derivative (slope) for the linear predictor in the longitudinal submodel(s); 
   shared random effects;
   no association structure (equivalent to fitting separate longitudinal and survival models). \cr
\cr
Estimation of the joint model is under a Bayesian framework using Markov chain 
Monte Carlo (MCMC) methods, with the underlying estimation implemented using 
Stan \url{http://mc-stan.org/}. A variety of prior distributions are available
for the regression coefficients, including normal, Cauchy and t distributions, 
as well as shrinkage priors. The prior distribution for the covariance matrix 
of the individual-level (or other clustering-level) random effects is based on 
a novel decomposition. All prior distributions are implemented via the 
\pkg{rstanarm} package; see \code{\link[rstanarm]{priors}} for details.
}
\references{
Stan Development Team. (2015). \emph{Stan Modeling Language Users Guide and
Reference Manual.} \url{http://mc-stan.org/documentation/}
}
\seealso{
\code{\link{stanjm-object}} and \code{\link{stanjm-methods}} 
  for details on the fitted model object returned by the \code{\link{stan_jm}}
  modelling function..
  
  \url{http://mc-stan.org/} for more information on the Stan C++ package used
  by \pkg{rstanjm} for model fitting.
}

