% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{summary.stanjm}
\alias{as.data.frame.summary.stanjm}
\alias{print.summary.stanjm}
\alias{summary.stanjm}
\title{Summary method for stanjm objects}
\usage{
\method{summary}{stanjm}(object, pars = c("long", "event"),
  regex_pars = NULL, probs = NULL, ..., digits = 1)

\method{print}{summary.stanjm}(x, digits = max(1, attr(x, "print.digits")),
  ...)

\method{as.data.frame}{summary.stanjm}(x, ...)
}
\arguments{
\item{object}{A fitted model object returned by the
\code{\link{stan_jm}} modelling function. See \code{\link{stanjm-object}}.}

\item{pars}{An optional character vector specifying a subset of parameters to
display. Parameters can be specified by name or several shortcuts can be 
used. Using \code{pars="long"} and/or \code{pars="event"} will display the 
parameter estimates for the longitudinal and event submodels, respectively,
but will not include coefficients for the subject-specific random effects. 
Using \code{pars="fixef"} will display parameter estimates for the fixed
effects from both the longitudinal and event submodels. 
Using \code{pars="beta"} will restrict the displayed parameters to 
only the regression coefficients (without the intercept). \code{"alpha"} 
can also be used as a shortcut for \code{"(Intercept)"}. The estimates for
the subject-specific random effects can be selected using \code{pars="b"}.
If \code{pars} is \code{NULL} all parameters are selected. See 
Examples.}

\item{regex_pars}{An optional character vector of \link[=grep]{regular 
expressions} to use for parameter selection. \code{regex_pars} can be used
in place of \code{pars} or in addition to \code{pars}. Currently, all
functions that accept a \code{regex_pars} argument ignore it for models fit
using optimization.}

\item{probs}{An optional numeric vector of probabilities passed to 
\code{\link[stats]{quantile}}, for calculating the quantiles of the 
posterior distribution for each parameter.}

\item{...}{Currently ignored.}

\item{digits}{Number of digits to use for formatting numbers when printing. 
When calling \code{summary}, the value of digits is stored as the 
\code{"print.digits"} attribute of the returned object.}

\item{x}{An object of class \code{"summary.stanjm"}.}
}
\value{
The \code{summary} method returns an object of class 
  \code{"summary.stanjm"}, which is a matrix of summary statistics and 
  diagnostics, with additional information stored as attributes.
  The \code{print} method for \code{summary.stanjm}
  objects is called for its side effect and just returns its input. The 
  \code{as.data.frame} method for \code{summary.stanjm} objects converts 
  the matrix to a data.frame, preserving row and column names but dropping  
  the \code{print}-related attributes.
}
\description{
The \code{summary} method for \code{\link[=stanjm-object]{stanjm}}
objects returns a summary of parameter estimates and MCMC convergence diagnostics 
(Monte Carlo error, effective sample size, Rhat) for a joint model fitted
using the \code{\link{stan_jm}} modelling function.
}
\examples{

}
\seealso{
\code{\link{print.stanjm}}, \code{\link{stanjm-methods}}
}

